fix review10: move classes packaging
