+++ extra/modules/greenbids-real-time-data/src/main/java/org/prebid/server/hooks/modules/greenbids/real/time/data/core/GreenbidsUserAgent.java
@@ -1,4 +1,4 @@
-package org.prebid.server.hooks.modules.greenbids.real.time.data.model.data;
+package org.prebid.server.hooks.modules.greenbids.real.time.data.core;
 
 import org.apache.commons.lang3.StringUtils;
 import ua_parser.Client;
